user all routes working

+++ apps/exp-backend/src/controllers/admin.controller.ts
@@ -7,9 +7,9 @@ export const getAccountDetails = async (req: Request, res: Response) => {
   try {
     const userId = req.userId;
 
-    const user = await prisma.user.findUnique({
+    const user = await prisma.admin.findUnique({
       where: { id: userId },
-      select: { id: true, username: true /* other fields */ },
+      select: { id: true, username: true, email: true, token: true },
     });
 
     if (!user) {
@@ -26,8 +26,7 @@ export const editAccountDetails = async (req: Request, res: Response) => {
   try {
     const userId = req.userId;
     const updatedData = req.body;
-
-    const updatedUser = await prisma.user.update({
+    const updatedUser = await prisma.admin.update({
       where: { id: userId },
       data: updatedData,
     });
@@ -43,14 +42,22 @@ export const deleteAccount = async (req: Request, res: Response) => {
   try {
     const userId = req.userId;
 
+    const existingUser = await prisma.admin.findUnique({
+      where: { id: userId },
+    });
+
+    if (!existingUser) {
+      return res.status(404).json({ error: "User not found" });
+    }
+
     await prisma.user.delete({
       where: { id: userId },
     });
 
-    res.status(204).send();
+    return res.status(204).send({ message: "Account deleted successfully" });
   } catch (error) {
     console.error("Error deleting account:", error);
-    res.status(500).json({ error: "Internal server error" });
+    return res.status(500).json({ error: "Internal server error" });
   }
 };
